Optimize INVERT-CHARACTER-MAP and fix bad final mapping of type 4 cmap
INVERT-CHARACTER-MAP was previously O(N^2), and could take ~1-5 seconds
when opening a file with many glyphs. Refactored to O(N), reducing
time spent on same files to ~0.005-0.008s.

Per microsoft docs, in a type 4 cmap: "...the final start code and
endCode values must be 0xFFFF. This segment need not contain any valid
mappings. (It can just map the single character code 0xFFFF to
missingGlyph)." At least one file maps it to index #xFFFF, which is
out of range. In that case, map it to 0.